Renamed test files for importers to make them consistent and clearer
